refactored engine tests to package
