+++ tests/engine/test_engine_scheduling.py
@@ -1,20 +1,8 @@
-"""Test the Event engine"""
 import pytest
-from tfin.engine import Engine, EngineState, Event, EventError, StopEngineError
 
+from tests.engine.conftest import event_factory, EmptyEvent
 
-@pytest.fixture
-def engine():
-    return Engine()
-
-
-class EmptyEvent(Event):
-    """An event to use in testing"""
-
-
-def event_factory(event_cls=EmptyEvent, timestep=0, name="Test Event", data={}):
-    return event_cls(timestep=timestep, name=name, data=data)
-
+from tfin.engine import EngineState, StopEngineError, Event, EventError
 
 def test_engine_scheduling(engine):
     """Tests that the engine schedules events sorted by timestep"""
@@ -37,20 +25,6 @@ def test_engine_schedule_bad_input(engine):
     engine.schedule("MyEvent")
     engine.schedule(42)
     assert len(engine.queue) == 0, "Bad data structure added to queue"
-
-
-def test_engine_init(engine):
-    """Test the engine is initialized correctly before setting up the environment"""
-    assert engine.is_state(EngineState.WAITING), "Engine state should init to WAITING"
-    assert "initialized" in engine.message.lower(), "Engine init message not set"
-
-
-def test_engine_str(engine):
-    """Test the custom str representation of the engine"""
-    for i in range(3):
-        engine.schedule(event_factory(timestep=i))
-
-    assert "3 events" in str(engine), str(engine)
 
 
 def test_engine_status_finished(engine):
